Fix for Italian numbers
There's a particular number scheme with certain Italian numbers. For geographic numbers
the maximum length is 10 digits. But this isn't the case when the NDC (such as 02) is following by
the number 1; in this case the maximum number length is 11. The changes in this pull request ensure that phony
is aware of this change and that numbers are split correctly.